Bump dependencies for HA 2025.2
The recent release of HA 2025.2 broke the ability to install the custom
integration that depends on this. This is because `uv export` currently
only supports exporting exact version dependencies, rather than
mirroring what is now in `pyproject.toml`. Until I find a better way to
export dependencies, I'll need to stay on top of HA upgrades made to
bleak and bleak-retry-connector.